BACKPORT: FROMGIT: jbd2: introduce jbd2_inode dirty range scoping
Currently both journal_submit_inode_data_buffers() and
journal_finish_inode_data_buffers() operate on the entire address space
of each of the inodes associated with a given journal entry.  The
consequence of this is that if we have an inode where we are constantly
appending dirty pages we can end up waiting for an indefinite amount of
time in journal_finish_inode_data_buffers() while we wait for all the
pages under writeback to be written out.

The easiest way to cause this type of workload is do just dd from
/dev/zero to a file until it fills the entire filesystem.  This can
cause journal_finish_inode_data_buffers() to wait for the duration of
the entire dd operation.

We can improve this situation by scoping each of the inode dirty ranges
associated with a given transaction.  We do this via the jbd2_inode
structure so that the scoping is contained within jbd2 and so that it
follows the lifetime and locking rules for that structure.

This allows us to limit the writeback & wait in
journal_submit_inode_data_buffers() and
journal_finish_inode_data_buffers() respectively to the dirty range for
a given struct jdb2_inode, keeping us from waiting forever if the inode
in question is still being appended to.

Signed-off-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Jan Kara <jack@suse.cz>
Cc: stable@vger.kernel.org

Regarding the backport, in recent kernels jbd2_journal_file_inode() has
been wrapped via jbd2_journal_inode_ranged_write() and
jbd2_journal_inode_ranged_wait().

BUG=b:130521177, chromium:959968, b:133470926, b:65528573, b:130032210
TEST=Multiple iteration of dd from /dev/zero to /mnt/stateful_partition,
filling the disk.  Also, copied multi-GiB files from USB sticks to the
Downloads folder via the GUI.  I did this on an eve.

(cherry picked from commit 6ba0e7dc64a5adcda2fbe65adc466891795d639e
 git://git.kernel.org/pub/scm/linux/kernel/git/tytso/ext4.git dev)
Change-Id: I4f0b9b8524ad966ae929c1e76e645e08ecdb4623
Reviewed-on: https://chromium-review.googlesource.com/1696101
Tested-by: Ross Zwisler <zwisler@google.com>
Commit-Ready: Ross Zwisler <zwisler@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yu Zhao <yuzhao@chromium.org>