CHROMIUM: arm64: dts: qcom: Fix DTS warnings due to thermal node
The unit-address must match the first address specified in the reg property of the node.
If the node has no reg property, the @ and unit-address can be omitted and the node-name alone differentiates
the node from other nodes at the same level in the tree.

BUG=b:137671902
Test=When the kernel sources are built, there should not be any warnings due to thermal node like:

has a unit name, but no reg property

Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>

Change-Id: I5696cba08946a3fa40ba68b6e44738739997a111
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1708535
Reviewed-by: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>