CHROMIUM: media: rockchip-vpu: Drop support for RK3288
RK3288 has never been tested on the chromeos-4.4 branch, it only
confuses anyone who finds this code and just makes it more difficult to
maintain the RK3399 code. The support was originally forward ported from
3.14 with the intention of having the code upstreamed, however the
upstream driver was completely redesigned and so there is no point in
keeping this old code in our codebase anymore. Let's drop it.

Note that some code referring to "rk3288" still remains as it is used
for RK3399 as well. I decided to rename only the code that had to be
moved from RK3288-specific files to reduce the amount of changes.

Also remove the encode-after-decode workaround as it is not needed on
RK3399.

BUG=none
TEST=emerge-scarlet chromeos-kernel-4_4

Change-Id: I703c9ada1a7158120ad17fe03710283313e42de0
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1738926
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>