CHROMIUM: WAR: misc: fastrpc: workaround DSP access issue
When the process waiting for the DSP is interrupted with a signal,
the so-called 'non-kernel' calls are also interrupted and therefore
return to the caller since they wait_interruptible on the callback.

Also, on calling device release, the DSP will callback in *sequence*
all those calls which it didnt complete due to them being
interrupted. And after those have been invoked, the DSP will send the
response to the release request.

It has been noticed that when the TEARDOWN request is interrupted and
then followed very closely by a RELEASE request, the DSP can stop
responding [1].

Adding a delay when calls have been interrupted seems to work around
this issue.

[1] either the DSP stops responding, or the endpoint stops working
since the driver does not receive any more callbacks on
fastrpc_rpmsg_callback.

BUG=b:138978210
TEST=label_image -a 1 in batch mode

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
Signed-off-by: Abhinav Asati <asatiabhi@codeaurora.org>
Change-Id: I0e3d6f63997c977bcf722fb098786fb9c63340ad
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>
Signed-off-by: Abhinav Asati <asatiabhi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1800597
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>