CHROMIUM: WAR: misc: fastrpc: fix double refcounting on dmabuf
dma buf refcount has to be done by the driver which is going to use the fd.
This driver already does refcount on the dmabuf fd if its actively using it
but also does an additional refcounting via extra ioctl.
this additional refcount can lead to memory leak in cases where the
applications fail to call the ioctl to decrement the refcount.

So remove this extra refcount in the ioctl.

More info of dma buf usage at drivers/dma-buf/dma-buf.c

BUG=b:138978210
TEST=label_image -a 1 in batch mode

Reported-by: Mayank Chopra <mchopra@quicinc.com>
Reported-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Abhinav Asati <asatiabhi@codeaurora.org>
Change-Id: I00d93493be881fb29f85423a6ce825cf5c2c5ac7
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>
Signed-off-by: Abhinav Asati <asatiabhi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1772924
Reviewed-by: Julan Hsu <julanhsu@google.com>
Commit-Queue: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>