CHROMIUM: arm64: dts: qcom: qcs404-mistral: Ensure buck switch frequency
For VSEL value of 0 the POR values for MODE is "Auto", in which for our
load typically would mean we're operating at a switching frequence of
hundreds of kHz. VSEL 1 on the other hand has the POR mode of Forced
PWM, which gives a switching frequence independent of load.

So switch to VSEL 1 as our primary operating mode, to have a
consistently high switching freqency.

BUG=b:141614863
TEST=Measure voltage levels of APPS_MODE_SEL and VREG_S3 during boot

Change-Id: Ifdb246478bdda484c15cbf1967bffd5146793088
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1848200
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>