CHROMIUM: bluetooth: Enable filter duplicated adv offload.
This change will reduce by an order of magnitude the number
of advertisements the system will need to process when an LE
device that distributed an IRK is paired but isn't connected.

This isn't a complete solution since the filter duplicate is
a best effort solution at best from the controller, so in very
noisy environements, it is possible that the controller will
be unable to filter duplicates and will send them all up to the
host.

BUG=b:143696105
TEST=Verify that advertisements duplicates are filtered after
pairing a bluetooth LE mouse and disconnecting it.

Change-Id: If1f038d3f6419ed23dfca5dbdcf783d3b9bb6b08
Signed-off-by: Alain Michaud <alainmichaud@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1899421
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Michael Sun <michaelfsun@google.com>