CHROMIUM: rk3288_dmc: Unregister cooling device when ddrfreq is disabled
When DDR frequency scaling is disabled we suspend the devfreq
device, which disables the periodic re-evaluation of the
device frequency. However suspending a devfreq device doesn't
inhabilitate cooling devices that may be associated with it.
In consequence a device may still receive frequency change
requests from the thermal subsystem, even though it is suspended.

In theory we could block requests in devfreq_set_target() for
suspended devices. This might be the right thing to do anyway,
however it is not sufficient. From the POV of the thermal subsystem
the cooling device would still be available, which doesn't represent
reality and might result in sub-optimal thermal decisions.

Unregister the rk3288_dmc cooling device when frequency scaling is
disabled and re-register it when it is enabled again.

BUG=chromium:1020174
TEST=on jerry:
  cooling device is removed when a HDMI display is plugged in and
  added again when the display is removed.

Change-Id: I2c8225a370b88269c8f48248db0f9e30bc208731
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1893535
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 3f12d76dc0f120c972b9fe448857778e2a84fc4d)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1898534
Tested-by: Douglas Anderson <dianders@chromium.org>